auth_service/requirements.txt
Flask==2.0.2
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.0
Flask-RESTful==0.3.9
email-validator==1.2.1
bcrypt==3.2.2
kafka-python==2.0.2
Flask-SocketIO==5.1.2
gevent==21.12.0
aiokafka==0.7.2